Add files via upload
Like a woman embellishing herself in grandeur and elegance, the body of a 

shopping mall is aesthetically fashioned from time to time. The linear structure of a 

shopping mall is modernist itself, from which a shopping mall departs, moving 

towards post-modernity by its narrative fashion. A shopping mall can be 

postmodernist by departing itself from the modernist linear structure. Telford Plaza, 

a local shopping mall in Hong Kong, consists of modernist and post-modernist 

elements, addressing the way of life of Hong Kong people through the spectacle of 

“the East Meets West”. The interplay of modernist and postmodernist elements evokes 

compelling sensations to strollers, with which “the East Meets West” resonates.